1) Merging of Yves work:   * Made tests compile (and work?) again.   * Fixed bug of first point of polygon not being drawn.   * Uppercased some class names.   * Fixed some OpenAir airspace files.   * Renamed Airspace::labelCoordinates to Airspace::labels   * Changed nbpoints for resolution from size_t to int because this didn't seem     to give any problems...   * ... and probably some other stuff I can't remember... ;-)

+++ src/Arc.cpp
@@ -118,14 +118,11 @@ void Arc::discretize( std::vector<Coordinate>& coords, double resolution ) const
   //double endAngle = getEndAngle() % 360;
   double startAngle = getStartAngle();
   double endAngle = getEndAngle();
-  //cout << "startAngle = " << startAngle << endl;
-  //cout << "endAngle= " << endAngle << endl;
 
   // Make sure that start angle is smaller than end angle.
   if (startAngle > endAngle)
     {
     startAngle = startAngle - 360;
-    //cout << "Changing startAngle to " << startAngle << endl;
     }
 
   double interval;
@@ -143,12 +140,13 @@ void Arc::discretize( std::vector<Coordinate>& coords, double resolution ) const
   double arcdegree_lon = lon.getArcDegree(lat);
 
   // Compute the number of points of the discretization.
-  size_t nbPoints = (size_t) ( 2 * pi * ( interval / 360.0 ) * radiusNM / resolution );
+  //size_t nbPoints = (size_t) ( 2 * pi * ( interval / 360.0 ) * radiusNM / resolution );
+  int nbPoints = (int) ( 2 * pi * ( interval / 360.0 ) * radiusNM / resolution );
 
   std::cout << "Arc -- nbPoints: " << nbPoints << std::endl;
 
   // Generate all points of the arc.
-  for (size_t i = 0; i < nbPoints; ++i)
+  for (int i = 0; i < nbPoints; ++i)
     {
     // 'Normal' geometric angles are COUNTERCLOCKWISE as follows:
     //
